differentiate between reader and writer failures
flow-indexer had a small bug introduced in 65ce12f46f443cd4731542125ca2312d4da8a222
where it couldn't tell the difference between an error reading a log
file and an error writing the filtered response to the client.  To fix
this, use a bufio writer which lets you get the result of the write
operation via Flush().  If the Flush failed, it means there was an error
writing to the client and the request should be immediately aborted.