Use default WriteBuffer size
Using a write buffer of 64M was causing flow-indexer to use 7G of ram
while indexing.  Changing this back to the default of 4M cuts the peak
usage in half.  I'm still not sure how a 64M buffer ends up turning into
7G of usage, but I think it has something to do with GC and leveldb
allocating a 64M buffer for each transaction.

This is a partial fix for #13