Pass original filenames directly to nfdump.
New versions of nfdump have implemented parallel reading, and can no
longer read from a pipe.

The auto decompression support isn't required for nfdump, so the
original filenames should be readable.

This is a bit of an ugly fix, but I need to think about what a better
backend interface should look like.  For most backends taking an
io.Reader is great, it's only nfdump that wants the original filename.
Changing all backends to take in an filename would make them a bit more
complicated, but possibly with the right helper function this will be a
non issue.